+++ src/Maker/Resources/FunctionalSmokeTest.provider.tpl.php
@@ -3,6 +3,7 @@ echo "<?php\n"; ?>
 
 namespace <?php echo $namespace; ?>;
 
+use PHPUnit\Framework\Attributes\DataProvider;
 <?php if ($with_dto): ?>
 use Pierstoval\SmokeTesting\FunctionalSmokeTester;
 use Pierstoval\SmokeTesting\FunctionalTestData;
@@ -15,29 +16,40 @@ class <?php echo $class_name; ?> extends WebTestCase
 <?php if ($with_dto): ?>
     use FunctionalSmokeTester;
 <?php endif; ?>
-<?php foreach ($routes as $route): ?>
 
-    public function testRoute<?php echo ucfirst(preg_replace_callback('~_([a-z0-9])~isUu', function($matches) {return strtoupper($matches[1]);}, $route['routeName'])).'WithMethod'.ucfirst(strtolower($route['httpMethod'])); ?>(): void
+    public static function provideRoutes(): \Generator
+    {
+    <?php foreach ($routes as $route): ?>
+        <?php if (str_starts_with($route['routePath'], '/_')) continue;
+        ?> yield '<?php echo $route['httpMethod']; ?> <?php echo $route['routePath']; ?>' => ['<?php echo $route['httpMethod']; ?>', '<?php echo $route['routePath']; ?>','<?php echo $route['routeName']; ?>'];
+    <?php endforeach; ?>
+    }
+
+    #[DataProvider('provideRoutes')]
+    public function testRoute(string $method, string $url, string $route): void
     {
 <?php if ($with_dto): ?>
         $this->runFunctionalTest(
-            FunctionalTestData::withUrl('<?php echo $route['routePath']; ?>')
-                ->withMethod('<?php echo $route['httpMethod']; ?>')
-                ->expectRouteName('<?php echo $route['routeName']; ?>')
-                ->appendCallableExpectation($this->assertStatusCodeLessThan500('<?php echo $route['httpMethod']; ?>', '<?php echo $route['routePath']; ?>'))
+            FunctionalTestData::withUrl($url)
+                ->withMethod($method)
+                ->expectRouteName($route)
+                ->appendCallableExpectation($this->assertStatusCodeLessThan500($method, $url))
         );
 <?php else: ?>
-        $client = static::createClient();
-        $client->request('GET', '/');
+    $client = static::createClient();
+    $client->request($method, $url);
 
-        static::assertLessThan(
-            500,
-            $client->getResponse()->getStatusCode(),
-            'Request "<?php echo $route['httpMethod'].' '.$route['routePath']; ?>" for route "<?php echo $route['routeName']; ?>" returned an internal error.',
-        );
+    static::assertLessThan(
+    500,
+    $client->getResponse()->getStatusCode(),
+    \sprintf(
+        'Request "%s %s" for route "%s" returned an internal error.',
+        $method, $url, $route
+    ),
+    );
 <?php endif; ?>
+
     }
-    <?php endforeach; ?>
 
 <?php if ($with_dto): ?>
     public function assertStatusCodeLessThan500(string $method, string $url): \Closure